+++ changelog.md
@@ -8,6 +8,7 @@
 - [FEATURE] Validations are now enabled by default for upsert.
 - [FEATURE] Preliminary support for `include.through.where`
 - [SECURITY/BUG] Fixed injection issue in direction param for order
+- [FEATURE/BUG] Raw queries always return all results (including affected rows etc). 
 
 # 2.0.0-rc7
 - [FEATURE] Throw an error if no where clause is given to `Model.destroy()`.